Type annotate requirements functions
### Summary

Type annotate requirement functions so that they are easier to understand, with syntax described here: https://www.typescriptlang.org/docs/handbook/type-checking-javascript-files.html
In this process, I found that the code is abusing JavaScript's feature of making `undefined` as falsy, so I tweaked the `RequirementFulfillment` object initializer to make them initialize to false so that it's more explicit.

### Test Plan

Add `@ts-check` to the top of the script section. Open VSCode. You will see that most the things are well-typed, except the first few imports. To make VSCode and TypeScript stop complaining about those imports, a more complete TS setup is necessary, but it should not be the focus of this diff. That's why `@ts-check` is not included in the diff.

### Notes

I think the requirement computation code should be moved to a separate file and make it a pure function, so that it's easier to reason and unit test.