Setup strict release workflows (#64)
### Summary

Setup a release workflow: such that commit pushed to `beta-release` branch will be deployed to prod directly.

This is a sensitive operation, so the `beta-release` branch is strictly branch-protected:

- Requires 3 reviews
- All CI checks must pass
- PR should be from `master -> beta-release`

### Test Plan

Code is adapted from samwise. See https://github.com/cornell-dti/samwise/pull/301
Something can only tested when we are doing deployment.

### Notes

We want to eventually change the branch name to `release` when the beta is over.