Start taking advantage of typescript setup (#68)
### Summary

#56 setups up typescript for this project. This diff starts the non-invasive refactoring effort to take advantage of the new setup.

I added type definitions for all requirements json in `src/requirements/types.ts`. Using this definition, we can also test the validity of the json schema via a typescript type checking run.

The type definition is not only useful for checking the json itself. It can also be used to the code that computes on the json. In `src/components/Requirements.vue`, I added a type import for requirements. Now autocompletion on requirement object will work perfectly with type shown. (You still need to annotate params as `Requirement` in jsdoc to enable the autocompletion.) To avoid too much disruption, I did not plan to migrate more code to TS in this diff.

Finally, I tweaked some ts and linter config so the correct setup can pass all the CI checks.

### Test Plan

- Run `npm run tsc`. It will succeed.
- Now break the requirement json. (e.g. change name to a int). Run `npm run tsc`. It will fail and produce a type error.