Compress decorated requirements json
### Summary

A large source of the big size of the json is caused by repetition of names. For example, if there is a requirement that all CS courses are eligible, then the word `CS` will be repeated for n * m times, where n = number of semesters, m =  number of CS courses.

To optimize this, we can store the classes in a tree structure rather than a flattened giant list. You can see the tree structure type definition in `src/requirements/types.ts`'s `EligibleCourses` type. With this structure, a lot of information are deduplicated! It also allows a much more efficient search, since we can avoid an O(n) search for each requirement.

The result of this optimization is amazing. The size reduced from 1.4M to less than 700K.