Support sub-requirements
### Summary

#### Rationale

Without sub-requirements, the nested include check will interfere with each other. For example, if a requirement requires `(A || B) && (C || D)`, the old infrastructure will lose the `and` and essentially only checks `(A || B) || (C || D)`.

#### Implementation

I updated all the checkers to either be a single function `(course: Course) => boolean` or an array of `(course: Course) => boolean`. An array of checker function represents the case that an requirement has sub-requirements, and each checker in the array correspond to one sub-requirement. As a result, the generated json now contains an array of `EligibleCourses` objects, where each object in array represents all courses that can satisfy one sub-requirement.

On the client side, the requirement checker is updated to use an array of courses. It groups satisfiable courses according to sub-requirement satisfiability. Then for each sub-requirement, as long there is one satisfiable course, we declare it done.

### Test Plan

Tested the original failing case of ['CS 2800', 'CS 2802']. Now it works!