Move all requirement code into one place (#83)
### Summary

This diff aims to do only one thing: keep all requirement code in one place.

In the last few commits, the requirement code can appear either in `requirement-generator` or `src/requirements`. It serves the experimentation of pre-computing requirements well. Now that we know the approach of pre-computing is right, it's some to cleanup the structure and consolidate some code.

There are already some code smell with the existing setup:

- `requirement-generator/types.ts` and `src/requirements/types.ts` are almost identical, but with minor differences
- Updating types are error-prone, because it's easy to forget updating it in another place, which causes broken code to pass TS checker.

Merging code into one place automatically solves the problem above.

Another issue of the current setup is that the current code is not self-contained. You need to waste 20min generating a `filtered-all-courses.json` before you can experimenting anything. In this commit, I finally decided to check this file into git, so that devs can have an easier life. It's around 37M, so it's not terribly big, and this file doesn't tend to change every often.

There are some caveats with merging the code together. I have documented these issues in respective files.

### Test Plan

Mostly moves file around. `npm run req-gen` still has the same behavior.