Strictify Requirements.vue (#84)
### Summary

To make further progress on making complex requirement map post-processing, it's important to strictify things a little bit so that we know the static types of those giant maps being passing around and manipulated. Therefore, I decide to convert `Requirements.vue` to be in TypeScript.

The conversion is mostly straightforward. If you read the diff, you will see that all I have added is type annotation. The only exception is that I changed `req.required` to `req.required || 0` to make TS happy, unless it will yell at us that `req.required` can be undefined.

### Test Plan

Played with the app. No behavior change.