Make requirement-functions.ts contains only pure functions (#85)
To help understanding the codebase, we don't want side effects that is related to UI rendering to entangle together with pure computation code. This diff unbreaks the entanglement and make everything in requirement-functions.ts pure functions.

This diff completes the effort necessary to cleanup the requirement codebase. After this, we can move on to implement support for filtering away illegally double-counted classes during post-process phase.