Migrate more files to TypeScript and remove broken user data fetch
### Summary

Trying to migrate some easy files to TypeScript. The goal is to increase some type coverage without increasing merge conflicts for other devs, so it's focused on files that have not been touched recently.

During the migration, I discovered the source of `missing or insufficient firebase permission error`. In `src/store.ts`, we are fetching a collection (`user`) that no longer exists. To prevent future confusion, I removed the code as well as reference to the user collection.

### Test Plan

Nothing breaks.