Retain more detailed course information in `RequirementFulfillment`
### Summary

Here begins the interesting refactoring part: in order to separate the phase of requirement computation into a requirement classification phase and a requirement postprocessing phase, the classification phase needs to output more information for the postprocessing phase to consume.

Currently, a lot of information is lost. For example, `coursesThatFulilledRequirement` field only stores the course code of the courses, not their full information. As a result, we have the computation of requirementMap and course fulfillment array entangled with each other.

This diff only changes the type of `courses` from code array (`string[]`) to nested array of full `CourseTaken` object (`CourseTaken[][]`). Now it can retain the structure of subrequirements, which can aid later computation.

### Test Plan

Everything still works.