Refactor types of requirement fulfillment (#88)
### Summary

It turns out that the previous few refactors are still not enough to implement double-counting detections and other potential post-processing functionality in a clean way.

This diff intends to address one of a few remaining problems. Currently, `RequirementFulfillment` object still has a bunch of field that's frontend related. The computation of these fields should not be entangled together with requirement satisfiability code.

This diff purges all uses of these bad practices for `MutableRequirementFulfillment`.

### Test Plan

Everything still works.