Convert `reqs.json` to typescript
### Summary

I'm preparing a double-counted course counting diff that might adds some extra fields to the json. To build confidence that things are correctly typed, we need to get rid of the last stronghold of giant untyped stuff: `reqs.json`

To be fair, that json is not completely untyped. It has been validated by `src/requirements/validate.js`. However, the validator is not strict and we may run the risk of increasing diverging away from it. If we can convert this to typescript, then we can take advantage of TS's native type checking ability to prevent the type definition from deviating from runtime behavior.

The diff size might look intimidatingly big, but most of the changes are due to the conversion from JSON to TS, but to generate the new TS file is fairly easy:

1. Rename `reqs.json` to `source-requirements-json.ts`
2. Prepend
```
/* eslint-disable max-len */
import { RequirementsJson } from './types';

const json: RequirementsJson =
```
to the front and append
```
;

export default json;
```
to the back of the new ts file.
3. Fix some type errors in the new ts file. (All the errors are because I forget to remove something like `search`, `includes` that should be gone in #81)

### Test Plan

Everything still works.