Provide helper higher order functions to make writing requirements more intuitive (#104)
### Summary

During course plan meeting, someone expressed concern that the new requirement setup is too hard to understand. This is a valid concern, since the new setup uses a lot of idea of functional programming and might not be friendly to beginners.

Reverting to the old setup is not an option since we don't want to lose the customization power. Therefore, we need to strike a balance between flexibility and readability.

The solution is actually not too hard: we can provide some higher order functions that abstract away the functional programming detail, to make you feel like you are writing the old nested array `includes`. Beginners can easily follow some examples without the need to really understand wha't happening under the hood.

### Test Plan

I use the new helper functions to rewrite CS requirements. It should be much easier to read now.

You can run `npm run req-gen` to see this change doesn't change the requirement computation result.