Merge requirement checkers with their metada
### Summary

Now that we have the requirement metadata in json and the checker is a more succient and more readable form, we can merge these together to improve developer experiences.

### Test Plan

Run `npm run req-gen`. The generated decorated json doesn't change, which shows that this diff completely preserves the requirement computation behavior.