Give a unique ID for each requirement
By giving each requirement a unique ID, we can test whether two requirement objects are the same much easily. It also gives us a way to avoid using array index as a way to find requirement in the frontend. It will also give me a space-efficient way to store user's choice on toggleable requirements, since I only need to use the requirement ID to tell which requirement is the choice associated with.