Factor out alerts from AppCourse object (#202)
We want to eventually make the data from firestore as the single source of truth. Therefore, we should aim to reduce the different between `FirestoreSemesterCourse` and `AppCourse`. In particular, we should avoid making `AppCourse` contain information that is not in the database (i.e. the stuff that can be computed on demand on the frontend).

The alerts information is one such example. In additional to introducing differences between frontend data type and db data type, this field inside `AppCourse` is also mutated in many different places. It's one of the reason that forces us to use the inefficient deep watch on semester array. Eliminating this mutable field from the object can also help make the flow cleaner.

This refactor diff computes a duplicate course id array instead. Then each course component can use this array to decide whether to display the duplicate warning. While coding for this diff, I also found some unused code, which I will annotate in the PR comments.