Cleanup NewCourse component by eliminating DOM manipulation (#259)
## Summary

The NewCourse component is also a mess. It's used in two different places:

- NewCourseModal for supporting the main functionality.
- OnboardingTransfer for selecting equivalent classes.

I first cleaned up the DOM manipulation by replacing that with standard vue state based approach. It turns out it's much simpler to write code in this way.

After the above cleanup, it becomes clear that OnboardingTransfer component is only using the new course component for selecting a course. All the extra requirement functionality are not needed. Therefore, I replaced the use of `newCourse` component to be `course-selector`. This change also unbreaks the functionality of adding transfer classes.

Then we are able to clean up a lot of if-else just to make the new-course component work for onboarding transfer.

## Test Plan

- check that OnboardingTransfer works. Try to add a transfer class.
- check that NewCourseModal works. Try to add a class inside a semester.

<img width="472" alt="Screen Shot 2021-02-03 at 23 22 02" src="https://user-images.githubusercontent.com/4290500/106844504-b859d080-6676-11eb-901a-0131c94f54fd.png">
<img width="503" alt="Screen Shot 2021-02-03 at 23 22 13" src="https://user-images.githubusercontent.com/4290500/106844505-b859d080-6676-11eb-8a1d-57da1767c2a9.png">
<img width="657" alt="Screen Shot 2021-02-03 at 23 22 31" src="https://user-images.githubusercontent.com/4290500/106844506-b859d080-6676-11eb-9907-662a8df25a58.png">