Factor out SelectedRequirementEditor (#264)
### Summary <!-- Required -->

This PR factors out SelectedRequirementEditor so that the logic of maintaining requirement editor is more clear and the code is more maintainable.

The key idea of this refactor is to make `SelectedRequirementEditor` completely stateless. It doesn't remember what's selected in its own data, but instead take it from the parent. The parent component (`NewCourse`) is the only component that remembers the selected state. In this way, the communication between `NewCourse` and `SelectedRequirementEditor` is much easier: `SelectedRequirementEditor` simply needs to tell the parent how to update the selected requirement list.

This PR also ensures that the child component compare requirement by id instead of name, so that we can guarantee that there will be no name collision.

I also renamed the BinaryButton to RequirementRadioSelectButton, because it's not always the case that the user is choosing between two choices. One example is MATH 4710 for engineering. It can be a CHEM replacement, an external spec or a free elective.

### Test Plan <!-- Required -->

Editor still works. (I added some mock data for `RequirementRadioSelectButton` since the choices are not generated yet)

<img width="472" alt="Screen Shot 2021-02-08 at 12 30 50" src="https://user-images.githubusercontent.com/4290500/107259068-8c3da700-6a0a-11eb-9c4e-696aab1e2aca.png">

### Notes <!-- Optional -->

I'm relatively sure that the current code for generating choices for `relatedRequirement` and `potentialRequirement` is wrong. It should inspect the graph instead of just looking at the requirement array. I will investigate the best way moving forward later. This PR serves as a foundation of making such changes easier in the future.

In the next PR, I plan to inline NewCourse.vue into NewCourseModal.vue in order to get rid of some ugly information passing using ref and some unnecessary indirection.