Remove subRequirementProgress field when req is fulfilled by credits (#310)
For all the requirements that are fulfilled by credits, it's always the case that there is only one checker produces eligible courses, and we want to count all the course credits from that list. Therefore, it's always `any-can-count`. (There are several that uses `every-course-needed`, but the credit summation code for `any-can-count` and `every-course-needed` is the same anyways...)

Thus, it doesn't make sense to store the field, and leaving it there might be very confusing. This diff removes them.