Change mechanism of disallowing dropping to fix drop cancel issue
## Summary

We don't need to rely on the move function's return value. Instead, we can use the [`group` option](https://github.com/SortableJS/Sortable#group-option) to disallow items being put into the list.

Surprisingly, this change also partially fixes the issue of [not dropping the req course into the semester still causing it to be added](https://www.notion.so/Cancel-Drag-Req-Course-8f5d03f44cb14cdf8576e13e19f3e21a). However, it's not a full fix, since you still need to drop the req course into the original req course slot. I don't believe a full fix is possible since it seems to be the fundamental assumption of the library that the card will stuck in the previous eligible slot until it finds a new slot.

## Test Plan

- Try to drop semester class into req courses and see it fail.
- Try to drag a req course into semester but not drop, then drag it back to the req course slot and drop. No change in semester data.